refactor: keep backendClient generic
